Mudança na classe EyeTrackingApp, melhor modularização
